Move hashIds to common
