+++ packages/web/src/pages/supporting-page/sagas.ts
@@ -12,6 +12,7 @@ import {
   getUserList,
   getUserIds
 } from 'common/store/user-list/supporting/selectors'
+import { decodeHashId, encodeHashId } from 'common/utils/hashIds'
 import { stringWeiToBN } from 'common/utils/wallet'
 import { createUserListProvider } from 'components/user-list/utils'
 import * as adapter from 'services/audius-api-client/ResponseAdapter'
@@ -19,7 +20,6 @@ import {
   fetchSupporting,
   SupportingResponse
 } from 'services/audius-backend/Tipping'
-import { decodeHashId, encodeHashId } from 'utils/route/hashIds'
 
 export const USER_LIST_TAG = 'SUPPORTING'
