+++ packages/web/src/utils/route/trackRouteParser.ts
@@ -1,9 +1,8 @@
 import { ID } from '@audius/common'
 import { matchPath } from 'react-router-dom'
 
+import { decodeHashId } from 'common/utils/hashIds'
 import { TRACK_ID_PAGE, TRACK_PAGE } from 'utils/route'
-
-import { decodeHashId } from './hashIds'
 
 export type TrackRouteParams =
   | { slug: string; trackId: null; handle: string }
